refactor: move private cores to subpackage
