+++ cachier/cores/mongo.py
@@ -20,7 +20,7 @@ try:
 except ImportError:  # pragma: no cover
     pass
 
-from .base_core import _BaseCore, RecalculationNeeded
+from .base import _BaseCore, RecalculationNeeded
 
 
 MONGO_SLEEP_DURATION_IN_SEC = 1
